# api/openapi/cloud-access-core.yaml
openapi: 3.1.0
info:
  title: Cloud Access Core API
  version: 0.9.0
  description: |
    REST surface for IAM, device onboarding, policy, and config center modules inside
    the 面试宝典 cloud-access-core service. JWT-protected admin endpoints use the IAM-issued
    access token; device endpoints use the device PASETO token. Deterministic error codes follow
    the scheme documented in docs/cloud-access-core-operations.md.
servers:
  - url: http://localhost:8080
    description: Default developer runtime (Makefile/Compose)
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: PASETO v4
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "40101"
        message:
          type: string
        request_id:
          type: string
          nullable: true
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in, refresh_expires_in, session_id, token_type]
      properties:
        access_token:
          type: string
        device_token:
          type: string
          description: Device-scoped PASETO token (missing for browser clients)
        refresh_token:
          type: string
        expires_in:
          type: integer
        refresh_expires_in:
          type: integer
        session_id:
          type: string
        token_type:
          type: string
          example: Bearer
        tenant_id:
          type: string
        user_id:
          type: string
        roles:
          type: array
          items:
            type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        client_kind:
          type: string
          description: web|desktop|api – influences token mix (JWT vs PASETO)
        device_fingerprint:
          type: string
          description: Optional identifier tied to the workstation
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    ConfigProfileUpsert:
      type: object
      required: [tenant_id, profile_name, sampling_rate, noise_gate]
      properties:
        tenant_id:
          type: string
          format: uuid
        profile_name:
          type: string
        sampling_rate:
          type: integer
        noise_gate:
          type: number
        detection:
          type: object
          additionalProperties: true
        model_presets:
          type: object
          additionalProperties: true
        endpoints:
          type: object
          additionalProperties: true
    ConfigProfileResponse:
      type: object
      properties:
        profile_name:
          type: string
        version:
          type: integer
        signature:
          type: string
        sampling_rate:
          type: integer
        noise_gate:
          type: number
        detection:
          type: object
          additionalProperties: true
        model_presets:
          type: object
          additionalProperties: true
        endpoints:
          type: object
          additionalProperties: true
        encrypted_profile:
          type: string
          description: Base64 AES-GCM payload (device fetch only)
...
paths:
  /auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT + refresh (and optional device token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Exchange a refresh token for a fresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Fresh token bundle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/logout:
    post:
      summary: Revoke a refresh token + active session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refresh token revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean
...
  /configs/profile:
    get:
      summary: Device fetches encrypted config bundle
      security:
        - DeviceToken: []
      parameters:
        - in: query
          required: true
          name: device_id
          schema:
            type: string
            format: uuid
        - in: query
          required: false
          name: profile
          schema:
            type: string
      responses:
        '200':
          description: Encrypted config profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProfileResponse'
  /configs/profiles:
    get:
      summary: Admin view of profile definition
      security:
        - BearerToken: []
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: profile
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Profile metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProfileResponse'
    put:
      summary: Upsert profile definition (admin only)
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigProfileUpsert'
      responses:
        '200':
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigProfileResponse'
...